refactor: readNewsAction
