+++ src/actions/news/read.ts
@@ -1,107 +1,24 @@
 'use server'
-import { newsArticleRealtionsIncluder } from '@/server/news/ConfigVars'
-import prisma from '@/prisma'
-import { createActionError, createPrismaActionError } from '@/actions/error'
 import type { ExpandedNewsArticle, SimpleNewsArticle } from '@/server/news/Types'
 import type { ActionReturn, ReadPageInput } from '@/actions/Types'
+import { readNews, readNewsCurrent, readOldNewsPage } from '@/server/news/read'
 
-export async function readOldNewsPage<const PageSize extends number>(
-    { page }: ReadPageInput<PageSize>
+export async function readOldNewsPageAction<const PageSize extends number>(
+    readPageImput: ReadPageInput<PageSize>
 ): Promise<ActionReturn<SimpleNewsArticle[]>> {
-    try {
-        const news = await prisma.newsArticle.findMany({
-            where: {
-                endDateTime: {
-                    lt: new Date(),
-                }
-            },
-            skip: page.page * page.pageSize,
-            take: page.pageSize,
-            orderBy: {
-                article: {
-                    createdAt: 'desc',
-                }
-            },
-            include: {
-                article: {
-                    include: {
-                        coverImage: {
-                            include: {
-                                image: true
-                            }
-                        }
-                    }
-                }
-            }
-        })
-        return {
-            success: true,
-            data: news.map(n => ({
-                ...n,
-                coverImage: n.article.coverImage.image
-            })),
-        }
-    } catch (error) {
-        return createPrismaActionError(error)
-    }
+    //TODO: only read news with right visibility
+    return await readOldNewsPage(readPageImput)
 }
 
-export async function readNewsCurrent(): Promise<ActionReturn<SimpleNewsArticle[]>> {
-    try {
-        const news = await prisma.newsArticle.findMany({
-            where: {
-                endDateTime: {
-                    gte: new Date(),
-                }
-            },
-            orderBy: {
-                article: {
-                    createdAt: 'desc',
-                },
-            },
-            include: {
-                article: {
-                    include: {
-                        coverImage: {
-                            include: {
-                                image: true
-                            }
-                        }
-                    }
-                }
-            }
-        })
-        return {
-            success: true,
-            data: news.map(n => ({
-                ...n,
-                coverImage: n.article.coverImage.image
-            }))
-        }
-    } catch (error) {
-        return createPrismaActionError(error)
-    }
+export async function readNewsCurrentAction(): Promise<ActionReturn<SimpleNewsArticle[]>> {
+    //TODO: only read news with right visibility
+    return await readNewsCurrent()
 }
 
-export async function readNewsByIdOrName(idOrName: number | {
+export async function readNewsAction(idOrName: number | {
     articleName: string
     order: number
 }): Promise<ActionReturn<ExpandedNewsArticle>> {
-    try {
-        const news = await prisma.newsArticle.findUnique({
-            where: typeof idOrName === 'number' ? {
-                id: idOrName
-            } : {
-                articleName_orderPublished: {
-                    articleName: idOrName.articleName,
-                    orderPublished: idOrName.order
-                }
-            },
-            include: newsArticleRealtionsIncluder
-        })
-        if (!news) return createActionError('NOT FOUND', `article ${idOrName} not found`)
-        return { success: true, data: news }
-    } catch (error) {
-        return createPrismaActionError(error)
-    }
+    //TODO: only read news if right visibility
+    return await readNews(idOrName)
 }
