+++ src/context/paging/OldNewsPaging.tsx
@@ -1,11 +1,11 @@
 'use client'
 import generatePagingProvider, { generatePagingContext } from './PagingGenerator'
-import { readOldNewsPage } from '@/actions/news/read'
+import { readOldNewsPageAction } from '@/actions/news/read'
 import type { ReadPageInput } from '@/actions/Types'
-import type { SimpleNewsArticle } from '@/actions/news/Types'
+import type { SimpleNewsArticle } from '@/server/news/Types'
 
 export type PageSizeOldNews = 20
-const fetcher = async (x: ReadPageInput<PageSizeOldNews>) => await readOldNewsPage(x)
+const fetcher = async (x: ReadPageInput<PageSizeOldNews>) => await readOldNewsPageAction(x)
 
 export const OldNewsPagingContext = generatePagingContext<SimpleNewsArticle, PageSizeOldNews>()
 const OldNewsPagingProvider = generatePagingProvider({ Context: OldNewsPagingContext, fetcher })
